+++ src/components/buyer/BuyerOrderList.tsx
@@ -1,44 +1,15 @@
-import {
-  TableContainer,
-  Table,
-  TableHead,
-  TableBody,
-  TableRow,
-  TableCell,
-  Paper,
-  Button,
-  Typography,
-} from '@mui/material';
-import { useEffect, useState } from 'react';
-
-import { api } from '../../api/api';
-import { IOrderItem } from '../../type';
+import { useState } from 'react';
 import { Link } from 'react-router-dom';
-import { ORDER_STATE } from '../../constants';
+import { Button, Typography } from '@mui/material';
+
+import { IOrderItem } from '../../type';
+import OrderListTable from './OrderListTable';
+import useSetOrderList from '../../hooks/useSetOrderList';
 
 export default function BuyerOrdeList() {
   const [orderList, setOrderList] = useState<IOrderItem[]>([]);
 
-  // 날짜 변환 함수
-  function formatDate(dateString: string) {
-    return new Intl.DateTimeFormat('ko-KR', {
-      year: 'numeric',
-      month: 'long',
-      day: 'numeric',
-    }).format(new Date(dateString));
-  }
-
-  useEffect(() => {
-    const getOrderProductInfo = async () => {
-      try {
-        const response = await api.getOrderProductInfo();
-        setOrderList(response.data.item);
-      } catch (error) {
-        console.log(error);
-      }
-    };
-    getOrderProductInfo();
-  }, []);
+  useSetOrderList(setOrderList);
 
   if (orderList.length === 0) {
     return (
@@ -57,60 +28,8 @@ export default function BuyerOrdeList() {
 
   return (
     <>
-      <TableContainer component={Paper}>
-        <Table aria-label="결제내역">
-          <TableHead>
-            <TableRow>
-              <TableCell align="center">
-                결제일
-                <br /> (주문번호)
-              </TableCell>
-              <TableCell align="center">상품명</TableCell>
-              <TableCell align="center">수량</TableCell>
-              <TableCell align="center">
-                총 결제금액 <br />
-                (배송비)
-              </TableCell>
-              <TableCell align="center">주문처리상태</TableCell>
-              <TableCell align="center"></TableCell>
-            </TableRow>
-          </TableHead>
-          <TableBody>
-            {orderList.map((rows) => (
-              <TableRow key={rows._id}>
-                <TableCell align="center">
-                  <Typography variant="body2" color="text.secondary">
-                    {formatDate(rows.createdAt)}
-                  </Typography>
-                  ({rows._id})
-                </TableCell>
-                <TableCell align="center">
-                  {rows.products[0].name} 포함 총 {rows.products.length}건
-                </TableCell>
-                <TableCell align="center">{rows.products.length}</TableCell>
-                <TableCell align="center">
-                  {rows.cost.total.toLocaleString()}원<br /> (
-                  {rows.cost.shippingFees.toLocaleString()}원)
-                </TableCell>
-                <TableCell align="center">
-                  {ORDER_STATE.codes
-                    .filter((state) => state.code === rows.state)
-                    .map((stateValue) => (
-                      <Typography key={stateValue.code} variant="body2">
-                        {stateValue.value}
-                      </Typography>
-                    ))}
-                </TableCell>
-                <TableCell align="center">
-                  <Button type="button" variant="outlined">
-                    별점평가
-                  </Button>
-                </TableCell>
-              </TableRow>
-            ))}
-          </TableBody>
-        </Table>
-      </TableContainer>
+      <Typography variant="h5">내 주문 내역</Typography>
+      <OrderListTable orderList={orderList} />
     </>
   );
 }
