+++ src/lib/formatDate.ts
@@ -1,0 +1,7 @@
+export default function formatDate(dateString: string) {
+  return new Intl.DateTimeFormat('ko-KR', {
+    year: 'numeric',
+    month: 'long',
+    day: 'numeric',
+  }).format(new Date(dateString));
+}
